% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents by year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer value representing a state in the FARS dataset (valid inputs: 1-56)}

\item{year}{integer value representing the year you want accidents plotted for}
}
\value{
A map plot of accidents in one state for one year
}
\description{
This function will create a map plot of accidents in a state for one year
}
\note{
State numbers outside the valid range (see above) will throw an error

States with no accidents to plot will produce a warning message
}
\examples{
\dontrun{fars_map_state(1, 2013)}
\dontrun{fars_map_state(56, 2015)}

}
