% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS files in}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string reference to a filename to read in or the actual filename you want to read in}
}
\value{
This function will return a dataset corresponding to the file you
requested to be loaded
}
\description{
This function will read in fars data files
}
\note{
Non-existent filenames will return an error
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read(make_filename(2013))}

}
