% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple years of FARS data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector input of the years you would like to import data for}
}
\value{
A list with each element containing a list of fatal accidents with the variables month and year
}
\description{
This function summarizes fatal accidents by month and year based on multiple FARS datafiles
}
\note{
Any year inputs that you do not have data for will generate a warning message.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
\dontrun{fars_read_years(c(2013,2014,2015))}

}
