% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector input of the years you would like to see summarized}
}
\value{
A table containing data on the number of accidents by year and month
}
\description{
This function will summarize the number of fatal accidents by month and year
}
\note{
Year inputs you do not have data for will return an error. If you submit a vector with
at least one valid input, then only a warning message will be generated.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014))}
\dontrun{fars_summarize_years(c(2013,2014,2015))}

}
