% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename reference}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric input for the years you want to read in FARS data for}
}
\value{
This function will return a string for a filename corresponding to the year input
}
\description{
This function will create a FARS filename reference for the year specified
}
\examples{
\dontrun{make_filename(2013)}

}
